+++ Makefile.am
@@ -764,8 +764,7 @@ test_apps += \
 	tests/test_pair_ipc \
 	tests/test_rebind_ipc \
 	tests/test_reqrep_ipc \
-	tests/test_use_fd_ipc \
-	tests/test_use_fd_tcp \
+	tests/test_use_fd \
 	tests/test_zmq_poll_fd \
 	tests/test_timeo \
 	tests/test_filter_ipc
@@ -796,15 +795,10 @@ tests_test_filter_ipc_SOURCES = tests/test_filter_ipc.cpp
 tests_test_filter_ipc_LDADD = src/libzmq.la ${UNITY_LIBS}
 tests_test_filter_ipc_CPPFLAGS = ${UNITY_CPPFLAGS}
 
-tests_test_use_fd_ipc_SOURCES = \
-	tests/test_use_fd_ipc.cpp \
+tests_test_use_fd_SOURCES = \
+	tests/test_use_fd.cpp \
 	tests/testutil.hpp
-tests_test_use_fd_ipc_LDADD = src/libzmq.la
-
-tests_test_use_fd_tcp_SOURCES = \
-	tests/test_use_fd_tcp.cpp \
-	tests/testutil.hpp
-tests_test_use_fd_tcp_LDADD = src/libzmq.la
+tests_test_use_fd_LDADD = src/libzmq.la
 
 tests_test_zmq_poll_fd_SOURCES = tests/test_zmq_poll_fd.cpp
 tests_test_zmq_poll_fd_LDADD = src/libzmq.la
